+++ starsky/starsky.feature.health/UpdateCheck/Models/ReleaseModel.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Text.Json.Serialization;
 
 namespace starsky.feature.health.UpdateCheck.Models
@@ -17,10 +18,20 @@ namespace starsky.feature.health.UpdateCheck.Models
 		[JsonPropertyName("draft")]
 		public bool Draft { get; set; }
 
+		private string _tagName = string.Empty;
+		
 		/// <summary>
-		/// tag_name
+		/// Should start with v
 		/// </summary>
 		[JsonPropertyName("tag_name")]
-		public string TagName { get; set; }
+		public string TagName {
+			get => _tagName;
+			set
+			{
+				if ( string.IsNullOrWhiteSpace(value)) return;
+				if ( !value.StartsWith("v") ) Console.WriteLine($"{_tagName} Should start with v");
+				_tagName = value;
+			} 
+		}
 	}
 }