+++ history.md
@@ -37,6 +37,7 @@ node starsky-tools/build-tools/app-version-update.js
 - [x]   (Changed) _Back-end_ UseRealtime (sockets) backend option changed to enable by default
 - [x]   (Fixed) _Front-end_ When updating files with realtime mode on, collection mode raws are shown after update
 - [x]   (Added) _Back-end_ API to check if current version is the latest on github releases
+- [x]   (Fixed) _Front-end_ Clientside check for latest version (click away for 4 days)
 
 # version 0.4.0-beta.1 - 2020-10-31
 _First release on Github Releases_